Opinion_entity_authorization

+++ StayStop/StayStop.BLL_EF/Service/OpinionService.cs
@@ -1,6 +1,9 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using StayStop.BLL.Authorization;
 using StayStop.BLL.Dtos.Opinion;
+using StayStop.BLL.Exceptions;
 using StayStop.BLL.IService;
 using StayStop.BLL_EF.Exceptions;
 using StayStop.DAL.Context;
@@ -17,11 +20,15 @@ namespace StayStop.BLL_EF.Service
     {
         private readonly StayStopDbContext _context;
         private readonly IMapper _mapper;
+        private readonly IUserContextService _userContextService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public OpinionService(StayStopDbContext context, IMapper mapper)
+        public OpinionService(StayStopDbContext context, IMapper mapper, IUserContextService userContextService, IAuthorizationService authorizationService)
         {
             _context = context;
             _mapper = mapper;
+            _userContextService = userContextService;
+            _authorizationService = authorizationService;
         }
 
         private Reservation GetReservationById(int reservationId)
@@ -44,7 +51,7 @@ namespace StayStop.BLL_EF.Service
             if (reservation.Opinion is not null) throw new ReservationAlreadyHasOpinion($"Reservation with an id {reservationId} already has opinion");
 
             var opinion = _mapper.Map<Opinion>(opinionDto);
-
+            opinion.AddedById = _userContextService.GetUserId;
             opinion.ReservationId = reservationId;
 
             reservation.Opinion = opinion;
@@ -61,7 +68,11 @@ namespace StayStop.BLL_EF.Service
             var reservation = GetReservationById(reservationId);
 
             if (reservation.Opinion is null) throw new InvalidOperationException($"Reservation with an id {reservationId} don't have opinion");
-
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, reservation.Opinion, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             reservation.Opinion = null;
 
             _context.Opinions.Remove(reservation.Opinion);
@@ -87,7 +98,11 @@ namespace StayStop.BLL_EF.Service
             var reservation = GetReservationById(reservationId);
 
             if (reservation.Opinion is null) throw new InvalidOperationException($"Reservation with an id {reservationId} don't have opinion");
-
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, reservation.Opinion, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbiddenException("Permission denied");
+            }
             _mapper.Map(opinionDto, reservation.Opinion);
 
             reservation.Opinion.ReservationId = reservationId;
